:package: package structure refactoring on skills
