+++ app/pages/index/index_skills/skillType.ts
