+++ app/routes/_index.tsx
@@ -1,12 +1,12 @@
 import Container from "@/components/common/Container";
-import IndexProfile from "@/pages/index/IndexProfile";
-import SkillsRow from "@/skills/SkillsRow";
+import IndexProfile from "@/pages/index/index_profile/IndexProfile";
+import IndexSkills from "@/pages/index/index_skills/IndexSkills";
 
 export default function Index() {
   return (
     <Container>
       <IndexProfile/>
-      <SkillsRow />
+      <IndexSkills />
     </Container>
   );
 }
